+++ isitfit/cli/core.py
@@ -14,10 +14,14 @@ import click
 
 from .. import isitfit_version
 
+from isitfit.utils import AwsProfileMan
+profile_man = AwsProfileMan()
+
 
 # For the --share-email "multiple options"
 # https://click.palletsprojects.com/en/7.x/options/#multiple-options
-
+# About --profile: docs at
+# https://click.palletsprojects.com/en/7.x/options/?highlight=prompt#values-from-environment-variables
 from isitfit.cli.click_descendents import isitfit_group
 @isitfit_group(invoke_without_command=False)
 @click.option('--debug', is_flag=True, help='Display more details to help with debugging')
@@ -26,8 +30,9 @@ from isitfit.cli.click_descendents import isitfit_group
 @click.option('--version', is_flag=True, help='DEPRECATED: use "isitfit version" instead', hidden=True)
 @click.option('--share-email', multiple=True, help='Share result to email address')
 @click.option('--skip-check-upgrade', is_flag=True, help='Skip step for checking for upgrade of isitfit')
+@click.option('--profile', type=str, default='default', callback=profile_man.validate_profile, prompt=profile_man.prompt(), help='Use a specific profile from your credential file.', envvar='AWS_PROFILE')
 @click.pass_context
-def cli_core(ctx, debug, verbose, optimize, version, share_email, skip_check_upgrade):
+def cli_core(ctx, debug, verbose, optimize, version, share_email, skip_check_upgrade, profile):
     # usage stats
     # https://docs.python.org/3.5/library/string.html#format-string-syntax
     from isitfit.utils import ping_matomo, b2l
@@ -109,6 +114,11 @@ def cli_core(ctx, debug, verbose, optimize, version, share_email, skip_check_upg
     # save `verbose` and `debug` for later tqdm
     ctx.obj['debug'] = debug
     ctx.obj['verbose'] = verbose
+
+    # set the profile in an env var so that boto3 picks it up automatically
+    import os
+    os.environ['AWS_PROFILE'] = profile
+